correct start.sh script

+++ test/start.sh
@@ -1,29 +1,30 @@
 #!/usr/bin/sh
 
-ganache-cli -i test > /dev/null &
-ganache_pid=$!
+function restart_ganache() {
+  echo "Restart ganache-cli..."
+  if [ $ganache_pid ]
+  then
+    echo "kill old ganache..."
+    kill -9 $ganache_pid
+  fi
 
-sleep 5
+  ganache-cli -i test > /dev/null &
+  ganache_pid=$!
+
+  sleep 5
+}
+
+restart_ganache
 
 echo "truffle test Bank..."
 truffle test test/testComplexBank.js test/token/* test/testOracle.js --network testBank
 
-kill -9 $ganache_pid
-
-ganache-cli -i test > /dev/null &
-ganache_pid=$!
-
-sleep 5 > /dev/null
+restart_ganache
 
 echo "truffle test Exchanger..."
 truffle test test/testLoans.js test/testComplexExchanger.js test/testDeposit.js --network testExchanger
 
-kill -9 $ganache_pid > /dev/null
-
-ganache-cli -i test > /dev/null &
-ganache_pid=$!
-
-sleep 5 > /dev/null
+restart_ganache
 
 echo "truffle test DAO..."
 truffle test test/testAssociation.js --network testDAO
